Change index_col=False to index_col=None
I don't understand why this is necessary, but the test error in
`test_concat_header` if I don't make this change. If you look at the
current stable documentation here

    https://pandas.pydata.org/pandas-docs/stable/generated/pandas.read_csv.html

you see the following written for `index_col`:

    index_col : int or sequence or False, default None
        Column to use as the row labels of the DataFrame. If a sequence
        is given, a MultiIndex is used. If you have a malformed file
        with delimiters at the end of each line, you might consider
        index_col=False to force pandas to _not_ use the first column as
        the index (row names)

So I'm not sure why index_col=None fixes the test, but it does. This
seems to directly contradict the documentation...